+++ Com/Test.js
@@ -2,15 +2,14 @@
  * Created by japjohal on 2019-01-13.
  */
 import React, { Component } from 'react';
-import { Alert, StyleSheet, Text, View, Button, TextInput, Image } from 'react-native';
+import {  StyleSheet, Text, View, Button, AsyncStorage } from 'react-native';
 import { connect } from 'react-redux'
-import {fetchPosts, newPost, newReduced  } from '../actions/postActions'
-import PropTypes from 'prop-types';
+import { fetchPosts, newPost, changeStatus } from '../actions/postActions'
 
 
-class Home extends  Component{
+class Test extends  Component{
     constructor(props){
-        super(props);
+        super(props)
         this.click = this.click.bind(this);
     }
     componentWillMount(){
@@ -18,16 +17,16 @@ class Home extends  Component{
     }
     click(){
         this.props.newPost("Jappadog");
+        console.log(this.props.loginStatus)
+        this.props.changeStatus();
+        AsyncStorage.removeItem("Status")
+            .then(resp => console.log("Async storage cleared: resp=> "+ resp))
     }
 
     render(){
-        // const returnVal = this.props.jap.map(post =>(
-        //         <Text>{post.title}</Text>
-        //     ))
-
         return(
-            <View>
-                <Text style={{fontSize:18, top:50, left:60}}> hello world</Text>
+            <View styles={{ color:'red'}}>
+                <Text style={{fontSize:18, top:50, left:60,color:'white'}}> hello world {this.props.loginStatus.toString()}</Text>
                 <Button
                     title="ClickME"
                     onPress={this.click}
@@ -37,9 +36,14 @@ class Home extends  Component{
     }
 }
 
+
+
+
 const mapStateToProps = state => ({
-    jap: state.posts.items,    // posts come from what you named the reducer in index. items comes from postreducer file
-    newItem: state.posts.item,
+    // jap: state.posts.items,    // posts come from what you named the reducer in index. items comes from postreducer file
+    // newItem: state.posts.item,
+    loginStatus: state.posts.loginStatus
 })
 
-export default connect(mapStateToProps, {fetchPosts, newPost})(Home)
+ export default connect(mapStateToProps, {fetchPosts, newPost, changeStatus})(Test)
+
