+++ actions/types.js
@@ -1,2 +1,3 @@
 export const FETCH_POSTS = "FETCH_POSTS";
 export const NEW_POST    = "NEW_POST";
+export const LOGIN_USER   = "LOGIN_USER";
