+++ package.json
@@ -7,14 +7,16 @@
     "test": "jest"
   },
   "dependencies": {
+    "i": "^0.3.6",
     "react": "16.6.3",
     "react-native": "0.58.3",
     "react-native-elements": "^1.0.0",
-  "react-native-vector-icons": "^6.2.0",
-  "react-redux": "^6.0.0",
-  "redux": "^4.0.1",
-  "redux-thunk": "^2.3.0"
-},
+    "react-native-vector-icons": "^6.2.0",
+    "react-navigation": "^3.1.5",
+    "react-redux": "^6.0.0",
+    "redux": "^4.0.1",
+    "redux-thunk": "^2.3.0"
+  },
   "devDependencies": {
     "babel-core": "^7.0.0-bridge.0",
     "babel-jest": "24.0.0",
